+++ test/functional/schema/schema.test.js
@@ -1,4 +1,5 @@
 const { cloneDeep } = require('lodash');
+const { makeExecutableSchema } = require('@graphql-tools/schema');
 const Query = require('../../../src/query/Query');
 const ASTSchema = require('../../../src/graphql/ast/Schema');
 const CoreSchema = require('../../../src/core/Schema');
@@ -39,14 +40,14 @@ describe('FNSchema', () => {
   });
 
   test('AST Base', () => {
-    const schema = new ASTSchema({ typeDefs: cloneDeep(baseGraphql) }).initialize();
+    const schema = new ASTSchema({ typeDefs: cloneDeep(baseGraphql) }, makeExecutableSchema).initialize();
     validate(schema);
     expect(schema.makeExecutableSchema()).toBeDefined();
     validate(schema);
   });
 
   test('Core Base', () => {
-    const schema = new CoreSchema({ typeDefs: cloneDeep(baseGraphql) }, stores).initialize();
+    const schema = new CoreSchema({ typeDefs: cloneDeep(baseGraphql) }, stores, makeExecutableSchema).initialize();
     validate(schema);
     expect(schema.makeExecutableSchema()).toBeDefined();
     validate(schema);
@@ -57,7 +58,7 @@ describe('FNSchema', () => {
   });
 
   test('getShape', () => {
-    const schema = new CoreSchema(schemaJS, stores).decorate();
+    const schema = new CoreSchema(schemaJS, stores, makeExecutableSchema).decorate();
     const artModel = schema.getModel('Art');
     expect(artModel).toBeDefined();