+++ test/graphql/SchemaDecorator.test.js
@@ -1,10 +1,11 @@
 const Path = require('path');
+const { makeExecutableSchema } = require('@graphql-tools/schema');
 const Schema = require('../../src/graphql/ast/Schema');
 const schemaFixtures = require('../fixtures/schema');
 
 describe('Schema', () => {
   test('mergeSchemaFromFiles', async () => {
-    const schema = new Schema();
+    const schema = new Schema({}, makeExecutableSchema);
     const modulePath = Path.join(__dirname, 'modules');
 
     //
@@ -31,7 +32,7 @@ describe('Schema', () => {
   });
 
   test('schema fixture', () => {
-    const schema = new Schema(schemaFixtures).decorate();
+    const schema = new Schema(schemaFixtures, makeExecutableSchema).decorate();
     expect(schema).toBeDefined();
 
     const xschema = schema.makeExecutableSchema();