+++ test/setup.js
@@ -1,3 +1,4 @@
+const { makeExecutableSchema } = require('@graphql-tools/schema');
 const { MongoMemoryReplSet } = require('mongodb-memory-server');
 const GraphQL = require('../src/core/GraphQL');
 const Schema = require('../src/core/Schema');
@@ -10,7 +11,7 @@ module.exports = async (context = {}) => {
   const mongoServer = await MongoMemoryReplSet.create({ replSet: { storageEngine: 'wiredTiger' } });
   const uri = mongoServer.getUri();
   stores.default.uri = uri;
-  const schema = new Schema(gqlSchema, stores);
+  const schema = new Schema(gqlSchema, stores, makeExecutableSchema);
   schema.decorate();
   const resolver = new Resolver(schema, context);
   context.autograph = { resolver };